requirements.txt
pandas
requests
matplotlib
numpy
seaborn
scikit-learn
mlflow
apache-airflow
dvc
dvc-s3
prometheus-client